requirement.txt
casadi
matplotlib
scipy
pybullet
tqdm
omegaconf
opencv-python
h5py
einops
torch==2.1.2
torchvision
tensorboard
icecream
robomimic==0.3.0
ipdb
wandb==0.15.0
pyzmq
robosuite
mujoco
pin
spatialmath-python
spdlog
gym
juliacall
dynamixel_sdk
ipython